1.6.5, Bloodhound and iFeed

+++ _app/core/addon.php
@@ -141,19 +141,62 @@ abstract class Addon
 
 
     /**
+     * Returns the config path for this add-on
+     * 
+     * @return string
+     */
+    public function getConfigPath()
+    {
+        return Config::getConfigPath() . '/add-ons/' . $this->addon_name . '/';
+    }
+
+
+    /**
      * Retrieves the config file for this Add-on
      *
      * @return mixed
      */
     public function getConfig()
     {
-        if (File::exists($file = Config::getConfigPath() . '/add-ons/' . $this->addon_name . '/' . $this->addon_name . '.yaml')) {
+        if (File::exists($file = $this->getConfigPath() . $this->addon_name . '.yaml')) {
             return YAML::parseFile($file);
         } elseif (File::exists($file = Config::getConfigPath() . '/add-ons/' . $this->addon_name . '.yaml')) {
             return YAML::parseFile($file);
         }
 
         return null;
+    }
+
+
+    /**
+     * Loads a given config file for this add-on
+     * 
+     * @param string  $path  Path to load relative to this add-on's config directory
+     * @param boolean  $log_error  Write an error message on fail?
+     * @param boolean  $throw_exception  Throw an exception on fail?
+     * @return array
+     * @throws Exception
+     */
+    public function loadConfigFile($path, $log_error=false, $throw_exception=false)
+    {
+        $path = trim($path);
+        $path .= (preg_match('/\.y[a]?ml$/i', $path, $matches)) ? '' : '.yaml';
+        
+        $full_path = $this->getConfigPath() . $path;
+        
+        if (!File::exists($full_path)) {
+            if ($log_error) {
+                $this->log->debug("Could not load config `" . $path . "`, file does not exist.");
+            }
+            
+            if ($throw_exception) {
+                throw new Exception("Could not load config `" . $path . "`, file does not exist.");
+            }
+            
+            return array();
+        }
+        
+        return YAML::parseFile($full_path);
     }
 
 
@@ -704,9 +747,13 @@ class ContextualCache extends ContextualObject
      */
     public function listAll($folder="")
     {
-        $path   = $this->contextualize($folder . "/");
+        $path = $this->contextualize($folder . "/");
+
         $finder = new Finder();
-        $files  = $finder->files()->in($path);
+        $files  = $finder->files()
+            ->in($path)
+            ->followLinks();
+
         $output = array();
 
         foreach ($files as $file) {
@@ -739,12 +786,20 @@ class ContextualCache extends ContextualObject
      * @return void
      */
     public function purgeOlderThan($seconds, $folder="")
-    {
+    {	    
         $this->isValidFilename($folder);
 
-        $path   = $this->contextualize($folder . "/");
+        $path  = $this->contextualize($folder . "/");
+	    
+	  if ($folder && !Folder::exists($path)) {
+		  return;
+	  }
+	    
         $finder = new Finder();
-        $files  = $finder->files()->in($path)->date("<= " . Date::format("F j, Y H:i:s", time() - $seconds));
+        $files  = $finder->files()
+            ->in($path)
+            ->date("<= " . Date::format("F j, Y H:i:s", time() - $seconds))
+            ->followLinks();
 
         foreach ($files as $file) {
             File::delete($file);
@@ -762,10 +817,17 @@ class ContextualCache extends ContextualObject
     public function purgeFromBefore($date, $folder="")
     {
         $this->isValidFilename($folder);
+        $path = $this->contextualize($folder . "/");
 
-        $path   = $this->contextualize($folder . "/");
+        if ($folder && !Folder::exists($path)) {
+            return;
+        }
+
         $finder = new Finder();
-        $files  = $finder->files()->in($path)->date("< " . Date::format("F j, Y H:i:s", $date));
+        $files  = $finder->files()
+            ->in($path)
+            ->date("< " . Date::format("F j, Y H:i:s", $date))
+            ->followLinks();
 
         foreach ($files as $file) {
             File::delete($file);
